chore: move DropdownBookItems to a dedicated file for reusability in BookDetails
- Refactored DropdownBookItems to make it reusable across different components
- Will be used in BookDetails for consistent behavior
