+++ frontend/src/components/books/BooksList.tsx
@@ -13,9 +13,9 @@ interface BooksListProps {
 const BooksList: React.FC<BooksListProps> = ({ readBooks, wantToRead, currentlyReading }) => {
     return (
         <Container style={{ marginTop: 20, marginBottom: 20, }}>
-            { readBooks.length > 0 && < BookShelf key='reading' title="Currently Reading" books={currentlyReading} />}
-            { wantToRead.length > 0 && < BookShelf key='to_read' title="Want to read" books={wantToRead} />}
-            { readBooks.length > 0 && < BookShelf key='read' title="Read" books={readBooks} />}
+            {readBooks.length > 0 && < BookShelf key='reading' title="Currently Reading" books={currentlyReading} />}
+            {wantToRead.length > 0 && < BookShelf key='to_read' title="Want to read" books={wantToRead} />}
+            {readBooks.length > 0 && < BookShelf key='read' title="Read" books={readBooks} />}
         </Container>
     );
 }
