+++ frontend/src/utils/BooksAPI.ts
@@ -4,6 +4,8 @@ const api = process.env.BOOKS_API_URL || 'http://127.0.0.1:5000'
 
 let token = localStorage.token;
 
+// TODO: this is a dummy token generated by JWT.io 
+//  and real token will be implemented with Auth0
 const header = 'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9';
 const payload = 'eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiZW1haWwiOiJqb2huLmRvZUBkb2UuY29tIiwiaWF0IjoxNTE2MjM5MDIyLCJwZXJtaXNzaW9ucyI6WyJib29rOnNhdmUiLCJib29rOmdldCIsImJvb2s6ZGVsZXRlIl19';
 const signature = 'sBr7N20iIwugvL0ix2m05oV1Op3Q1oAdhW05KhUbZWA';
@@ -36,7 +38,7 @@ export const getBook = (id: string) =>
     .then((data) => data.book);
 
 export const update = (id: string, shelf: string) =>
-  fetch(`${api}/books/${id}`, {
+  fetch(`${api}/book/${id}`, {
     method: "PATCH",
     headers: {
       ...headers,
